requirements.txt
asgiref==3.2.7
autopep8==1.5.3
beautifulsoup4==4.9.1
Django==3.0.7
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==2.0.1
importlib-metadata==1.6.1
pycodestyle==2.6.0
python-decouple==3.3
pytz==2020.1
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
zipp==3.1.0
